% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{chessPI.pull}
\alias{chessPI.pull}
\title{Download and output games as a dataframe}
\usage{
chessPI.pull(username = "Finished_Milk", range = "all")
}
\arguments{
\item{username}{The Chess.com username}

\item{range}{the date provided by the archive function. If left blank, ALL games for the user will be downloaded.}
}
\value{
A dataset containing the games.
}
\description{
Pull your games from Chess.com
}
\examples{
games <- chessPI.pull("Finished_Milk","2024/01")

}
